Context management
The context will be used to analyze text in a different manner : the
code isn't formed in the same way in a declaration section than in the
behavior section.

The VHDL parser manage now a list of VP_Word instead of a list of Word.
VP_Word class know if the word is a VHDL keyword.

Add a header file describes the VHDL syntax.